refactor(scalar-arrays): expose raw expression AST on ResolvedAttributeArg (slice 2 N1)
Replace the decoded ResolvedExpr union with the parser ExpressionAst node
directly on ResolvedAttributeArg, per code-owner review. List-default parsing
narrows via the AST classes (ArrayLiteralAst/StringLiteralExprAst/...) instead
of a bespoke decoded shape. The stringified value is kept, as it is still
consumed by scalar/function-call/relation-name/authoring-arg parsing.
