+++ packages/1-framework/2-authoring/psl-parser/src/exports/index.ts
@@ -54,7 +54,6 @@ export type {
   NamespaceSymbol,
   ResolvedAttribute,
   ResolvedAttributeArg,
-  ResolvedExpr,
   ResolvedNamedTypeBinding,
   ResolvedTypeConstructorCall,
   ScalarSymbol,
