+++ packages/1-framework/2-authoring/psl-parser/src/symbol-table.ts
@@ -27,7 +27,6 @@ import type { SyntaxNode } from './syntax/red';
 export type {
   ResolvedAttribute,
   ResolvedAttributeArg,
-  ResolvedExpr,
   ResolvedTypeConstructorCall,
 } from './resolve';
