+++ packages/2-sql/2-authoring/contract-psl/src/psl-attribute-parsing.ts
@@ -1,7 +1,8 @@
 import type { ContractSourceDiagnostic } from '@prisma-next/config/config-types';
 import type { ControlPolicy } from '@prisma-next/contract/types';
-import type { PslSpan, ResolvedAttribute, ResolvedExpr } from '@prisma-next/psl-parser';
+import type { PslSpan, ResolvedAttribute } from '@prisma-next/psl-parser';
 import { parseQuotedStringLiteral } from '@prisma-next/psl-parser';
+import type { ExpressionAst } from '@prisma-next/psl-parser/syntax';
 
 export { parseQuotedStringLiteral };
 
@@ -32,7 +33,7 @@ export function getNamedArgument(attribute: ResolvedAttribute, name: string): st
 export function getPositionalArgumentEntry(
   attribute: ResolvedAttribute,
   index = 0,
-): { value: string; expression?: ResolvedExpr; span: PslSpan } | undefined {
+): { value: string; expression?: ExpressionAst; span: PslSpan } | undefined {
   const entries = attribute.args.filter((arg) => arg.kind === 'positional');
   const entry = entries[index];
   if (!entry || entry.kind !== 'positional') {
